requirements.txt
bayesian-optimization
beautifulsoup4
scikit-learn
xgboost
matplotlib
numpy
pandas
requests
seaborn
lxml
psycopg2-binary
torch>=2.2.0
joblib
